filter by audit stage
An api server request is captured twice in the apiserver audit log.
It is known as stage, the first entry is 'RequestReceived', it is
added as soon as the request is received and does not have any
response status. The second entry is 'ResponseComplete' and it is
written as soon as the request is completed with a response.

The stage filter adds ability to filter by these stages. The default
behavior is include all stages which maintains backward compatibility.